+++ local.sh
@@ -13,10 +13,12 @@ python3 -m pip install --user .
 ufotest init --update
 
 #ufotest init -f
-#ufotest --mock test single_frame
-ufotest --mock test --suite "routine"
+ufotest --verbose --mock test scripts_syntax
+#ufotest --verbose --mock test --suite "full"
 
 # Now we start the web interface server
 ufotest --mock ci serve
+
+